Fix Windows build failure caused by missing dependency. (#969)
Msys2's libcurl package was recently updated to enable HTTP/3
support which means it now also depends on libnghttp3.
genDllVersions.sh was updated to include this new dependency.

The Windows installer build action was also changed to upload its
build artifacts before it tests to make sure the plugins load. This
makes it possible to look at the generated artifacts for missing
dependencies if the plugin load verification fails.