RB-2.5 => RB-2.6 Merge fixes (#989)
* Update README to reflect 2.5 release

* Remove Bountysource from Readme

Bountysource appears to be dead, so it should be removed

* macports update

* Update README.md

* update MacPorts

* Update compiler-common.sh

* Debug build fix

* CMake: Debug build type fix

* Comment out check for debug flag build type. (#972)

* Fix windows pacman repo build. (#968)

Msys2 updated gcc in the last few days which has caused the package
repo build to fail. Here are some minor fixes to make the build work
again.

- Add fmt to openimageio's dependency list since the version included
  with the openimageio source now appears to cause an internal compiler
  error.

- Added an include for math.h to in a libsox test file to make the
  compiler stop generating an error about fabs() not being defined.

- Fix type mismatch and missing declaration errors in ImageMagick.

- Updated mingw-w64-dump_syms gitcommit to point to a version that
  fixes a missing header build error.

* Fix Windows build failure caused by missing dependency. (#969)

Msys2's libcurl package was recently updated to enable HTTP/3
support which means it now also depends on libnghttp3.
genDllVersions.sh was updated to include this new dependency.

The Windows installer build action was also changed to upload its
build artifacts before it tests to make sure the plugins load. This
makes it possible to look at the generated artifacts for missing
dependencies if the plugin load verification fails.

* Fix Windows debug cmake build. (#973)

- Force Shiboken to use detected python libraries instead of
  using its own library detection logic.
- Specify -O compiler option for debug builds like the qmake build.
- Remove unconditional NDEBUG define added by the Shiboken cmake
  helper so that asserts work correctly on debug builds.

* Add logic to capture/restore existing GL context.

This fixes asserts in debug builds where Qt's GL context is set when
a ScopedGLContext is created. This can happen when we have a project
open and the user tries to open another project. When the event
callback is dispatched, the Qt GL context is still set. The fix
justs removes the assert and adds logic to save and restore any
existing GL context that is set.

* MacPorts update

* Fix rotopaint overlay 2.5 (#986)

* RotoPaint: always draw the overlay when the tool is part of the overlay

* tools/MacPorts: remove .rej files

---------

Co-authored-by: Marcel Ruegenberg <gh2@picodesign.de>
Co-authored-by: Frédéric Devernay <devernay@users.noreply.github.com>
Co-authored-by: kafxa <148536772+kafxa@users.noreply.github.com>
Co-authored-by: Aaron Colwell <300262+acolwell@users.noreply.github.com>